AutoMLOps/components/component_base/requirements.txt
db_dtypes
fsspec
gcsfs
google-cloud-aiplatform
google-cloud-appengine-logging
google-cloud-audit-log
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-bigtable
google-cloud-core
google-cloud-dataproc
google-cloud-datastore
google-cloud-dlp
google-cloud-firestore
google-cloud-kms
google-cloud-language
google-cloud-logging
google-cloud-monitoring
google-cloud-notebooks
google-cloud-pipeline-components
google-cloud-pubsub
google-cloud-pubsublite
google-cloud-recommendations-ai
google-cloud-resource-manager
google-cloud-scheduler
google-cloud-spanner
google-cloud-speech
google-cloud-storage
google-cloud-tasks
google-cloud-translate
google-cloud-videointelligence
google-cloud-vision
kfp>=2.0.0
kfp_pipeline_spec
pandas
protobuf
pyarrow
scikit_learn
xgboost
